pyproject.toml
[tool.poetry]
name = "golem-backend"
version = "0.1.0"
description = ""
authors = ["Erick Medeiros <erick.medeiros.104@ufrn.edu.br>"]
readme = "README.md"
packages = [{include = "golem_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
django-model-utils = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
